+++ packages/framework/src/Services/AssetService.php
@@ -22,14 +22,6 @@ class AssetService implements AssetServiceContract
         return $this->version;
     }
 
-    /**
-     * @deprecated v0.50.x - Use cdnLink() instead.
-     */
-    public function stylePath(): string
-    {
-        return $this->constructCdnPath('hyde.css');
-    }
-
     public function constructCdnPath(string $file): string
     {
         return 'https://cdn.jsdelivr.net/npm/hydefront@'.$this->version().'/dist/'.$file;