+++ RELEASE_NOTES.md
@@ -20,6 +20,7 @@ This serves two purposes:
 
 ### Removed
 - Removed deprecated legacy Compiler class from the Hyde Realtime Compiler
+- Removed deprecated stylePath method from AssetService (deprecated in v0.50)
 
 ### Fixed
 - for any bug fixes.